refactored code and merged all tokenizer functionality into less files
